requirements.txt
tqdm==4.45.0
python-ldap==3.3.1
requests==2.25.1